Creates new core list component
