+++ src/components/dynamic-lists/beagle-grid-view/beagle-grid-view.component.html
@@ -14,9 +14,18 @@
   * limitations under the License. 
 -->
 
-<div 
-  class="beagle-grid-view"
-  [ngStyle]="{'grid-template-columns': getColumnsQuantityStyle()}"
-  [beagleAccessibility]="accessibility">
-  <ng-content></ng-content>
-</div>
+<beagle-dynamic-list 
+  type='GRID'
+  [numColumns] = numColumns
+  [dataSource] = dataSource
+  [iteratorName] = iteratorName
+  [template] = template
+  [onInit] = onInit
+  [onScrollEnd] = onScrollEnd
+  [scrollEndThreshold] = scrollEndThreshold
+  [useParentScroll] = useParentScroll
+  [key] = key
+  [__suffix__]= __suffix__
+  [isScrollIndicatorVisible] = isScrollIndicatorVisible
+  [beagleComponent] = this>
+</beagle-dynamic-list>
