+++ src/components/dynamic-lists/beagle-list-view/beagle-list-view.component.html
@@ -14,11 +14,19 @@
   * limitations under the License. 
 -->
 
-<div 
-  class="beagle-list-view" 
-  [ngClass]="direction" 
-  [attr.aria-colcount]="direction === 'VERTICAL' ? 1 : dataSource.length"
-  [attr.aria-rowcount]="direction === 'VERTICAL' ? this.dataSource.length : 1"
-  [beagleAccessibility]="accessibility">
-  <ng-content></ng-content>
-</div>
+<beagle-dynamic-list 
+  type='LIST'
+  [direction] = direction
+  [dataSource] = dataSource
+  [iteratorName] = iteratorName
+  [template] = template
+  [onInit] = onInit
+  [onScrollEnd] = onScrollEnd
+  [scrollEndThreshold] = scrollEndThreshold
+  [useParentScroll] = useParentScroll
+  [key] = key
+  [__suffix__]= __suffix__
+  [isScrollIndicatorVisible] = isScrollIndicatorVisible
+  [beagleComponent] = this>
+  
+</beagle-dynamic-list>
