+++ src/components/dynamic-lists/dynamic-list.component.html
@@ -1,4 +1,4 @@
-/*
+<!-- 
   * Copyright 2020 ZUP IT SERVICOS EM TECNOLOGIA E INOVACAO SA
   *
   * Licensed under the Apache License, Version 2.0 (the "License");
@@ -11,41 +11,14 @@
   * distributed under the License is distributed on an "AS IS" BASIS,
   * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
   * See the License for the specific language governing permissions and
-  * limitations under the License.
-*/
+  * limitations under the License. 
+-->
 
-beagle-list-view {
-  display: block;
-
-  .beagle-list-view {
-    display: flex;
-
-    &.VERTICAL {
-      flex-direction: column;
-    }
-
-    &.HORIZONTAL {
-      flex-direction: row;
-    }
-  }
-
-  &.hasScroll {
-    overflow: auto;
-
-    &.HORIZONTAL {
-      width: 100%;
-    }
-
-    &.VERTICAL {
-      height: 100%;
-    }
-  }
-}
-
-.hide-scrollbar {
-  scrollbar-width: none; /* Firefox 64 */
-  -ms-overflow-style: none; /* Internet Explorer 11 */
-  &::-webkit-scrollbar { /** WebKit */
-    display: none;
-  }
-}
+<div 
+  [ngClass]= getClassForType()
+  [ngStyle]="{'grid-template-columns': type === 'GRID' && getColumnsQuantityStyle()}" 
+  [attr.aria-colcount]="type === 'LIST' && (direction === 'VERTICAL' ? 1 : dataSource.length)"
+  [attr.aria-rowcount]="type === 'LIST' && (direction === 'VERTICAL' ? this.dataSource.length : 1)"
+  [beagleAccessibility]="accessibility">
+  <ng-content></ng-content>
+</div>
